Add two-dimensional vector
Add a generic two-dimensional vector which supports numeric types.

One, caveat is that it actually only supports integers, floats, and
complex numbers right now. It would be preferable to use a generic `T`
with a bound of `numbers.Number` (`T = TypeVar('T', bound=Number)`, but
mypy doesn't seem to be able to understand that the builtin number types
are `number.Number`s.