Change integer size 2d to be a vector 2
Change the integer size 2D to be an integer vector in two dimensions.

Still enable getting and setting of width and height even though the
vector2 uses x and y.

We should make a mechanism for aliasing instance variables. Maybe using
a decorator? We should also work on automatically generated reprs
and make sure that if the alias is preffered over the orignal name, then
the alias is used in the repr.