Change to comparing lists not strings
In the Illud print test, compare lists of the captured output and the
expected output. The reason for doing this is that if the direct
string are compared, and they do not match, then they will printed by
pytest. When pytest prints them, it doesn't escape them and so
consequently the output of pytest is garbled severely by the ANSI
escape codes.