Refactor cursor drawing
Seperate cursor drawing from window drawing. This does mean that we end
up writing more characters to stdout right now, but it has some
advantages. We no longer violate the single responsibility principle and
it makes the code easier to understand. Eventually, we can introduce a
layer between the terminal representation and the actual terminal which
handles all the drawing for us and gives us a much better interface than
printing directly.