Change Illud state initialization
Change Illud state to be initialized with sub-state directly.

There is a bit of a problem here: The state is a graph, but we are
treating it like a tree by using classes. For now, let's allow
passing each node in the graph so that the edge formation can be done
prior to initialization of Illud state. Eventually, it would be nice
to be able to construct the state from a dictionary (maybe even typed)
with the ability to use references to form edges.