Add equality operator to the standard output
Add an equality operator to the standard output so that we can more
easily check that other objects which contain it (such as the canvas)
can be compared for equality with the Seligimus decorators. The
decorator for comparing instance attributes doesn't yet support
excluding specific attributes, otherwise we could use that to ignore the
standard output attribute.