Add expand selection right
Add the ability to expand the selection to the right. If we want to be
able to manipulate the start and end of the selection like the cursor
can be manipulated, we are going to want to extract a lot of the common
code out of the cursor to reduce duplication. For now, let's just add
this one simple manipulation.