Fix window drawing to use position
Fix window drawing to use the window position. I took an approach here
of rewriting the entire function because I thought I would be able to
clean it up in doing so, but also because I thought the function before
was too hard to read and understand. So, I thought it would be faster to
rewrite. I think I was right on that account, but that I was wrong that
it would end up cleaning up. The function is still gross. I wonder if it
is in someway limitted in how clean it can become by the underlying
objects it acts on. Does the buffer need to use lines for this to have
the capacity for being clean? How much of the cleanliness of a system is
capped by it's components?