Add canvas fill
Add the ability to fill a canvas with a given character. I'm not sure
how I feel about this acting as both an operation on the canvas and also
returning a canvas. It is more useful in test cases if it retuns a
canvas but might lead to confusion in the future. Maybe an optional fill
parameter in the constructor would be a better solution?