Add mode to status bar
Add the name of the current mode to the status bar. Long term we need
to come up with a better way of managing state display. For example,
there could be a mechanism for registing callbacks when certain pieces
of state change. So something would be called when the mode changes and
that would change the status bar. Status bar layout also probably needs
some work. It might make sense to have several windows for it which are
grouped.